+++ ChangeLog.md
@@ -1,5 +1,6 @@
 0.26.0
 - have test callback methods `done`, `skip` and `fail` be bound functions
+- fail test on all thrown errors, even if the error value is falsy or null
 
 0.25.3
 - upgrade qmock to latest fixed version